Added folder'fixture' and moved files to folder
